pyproject.toml
[tool.poetry]
name = "polybar-gmail"
version = "0.1.0"
description = "A Polybar module to show unread messages from Gmail"
authors = ["Vyacheslav Konovalov <vyachkonovalov@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^1.8.2"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
